+++ src/cmps/stepfour/step-four-art.jsx
@@ -1,17 +1,24 @@
 import React, { useContext, useState } from 'react';
 
 import { Formik, Form, Field } from "formik";
-import StepThreeAssociations from './step-three-associations';
-import StepThreekids from './step-three-kids';
-import { SiteContext } from '../store/context';
+import StepThreeAssociations from '../stepthree/step-three-associations';
+import StepThreekids from '../stepthree/step-three-kids';
+import { SiteContext } from '../../store/context';
 
 
 export default function StepFourArt({ formikProps }) {
 	const [artCount, setArtCount] = useState(0)
 
 	const {
+		data,
+		setData,
 		moveNextStep
-	} = useContext(SiteContext)
+	} = useContext(SiteContext);
+
+	const handleSubmit = (values, actions) => {
+		setData(prev => ({ ...prev, values }))
+		moveNextStep(true);
+	}
 
 	// TODO: Mutating state
 	const onRemoveArt = (i) => {
@@ -112,28 +119,42 @@ export default function StepFourArt({ formikProps }) {
 
 	return (
 		<>
-			<div role="group">
-				<h4 style={{ margin: '10px 0' }}>האם קיימים ברשותך אומנות?</h4>
-				<div className="status-group flex space-between input-btn">
-					<label className={`${formikProps.values.art === "לא" ? 'active' : ''}`}>
-						<Field type="radio" name="art" value="לא" />
-						לא
-					</label>
-					<label className={`${formikProps.values.art === "כן" ? 'active' : ''}`}>
-						<Field type="radio" name="art" value="כן" />
-						כן
-					</label>
-				</div>
-			</div>
-			{formikProps.values.art === 'כן' &&
-				<div className="input-container-formik real-estate-container direction-ltr" >
-					{renderArtsForm(formikProps)}
-				</div>
-			}
+			<Formik
+				//validationSchema={validationSchema}
+				initialValues={data}
+				onSubmit={handleSubmit}
+			>
+				{({ values }) => {
+					return (
+						<Form>
+							<div role="group">
+								<h4 style={{ margin: '10px 0' }}>האם קיימים ברשותך אומנות?</h4>
+								<div className="status-group flex space-between input-btn">
+									<label className={`${formikProps.values.art === "לא" ? 'active' : ''}`}>
+										<Field type="radio" name="art" value="לא" />
+										לא
+									</label>
+									<label className={`${formikProps.values.art === "כן" ? 'active' : ''}`}>
+										<Field type="radio" name="art" value="כן" />
+										כן
+									</label>
+								</div>
+							</div>
+							{formikProps.values.art === 'כן' &&
+								<div className="input-container-formik real-estate-container direction-ltr" >
+									{renderArtsForm(formikProps)}
+								</div>
+							}
 
 
 
-			<button onClick={() => moveNextStep(true)}>המשך</button>
+							<button type="submit">המשך</button>
+						</Form>
+					)
+				}
+				}
+			</Formik>
+
 		</>
 	)
 }
