+++ src/cmps/stepfour/step-four-tools.jsx
@@ -1,12 +1,12 @@
 import React, { useContext, useState } from 'react';
 
 import { Formik, Form, Field } from "formik";
-import StepThreeAssociations from './step-three-associations';
-import StepThreekids from './step-three-kids';
-import { SiteContext } from '../store/context';
+import StepThreeAssociations from '../stepthree/step-three-associations';
+import StepThreekids from '../stepthree/step-three-kids';
+import { SiteContext } from '../../store/context';
 
 
-export default function StepFourTool({ formikProps }) {
+export default function StepFourTool() {
 	const [toolCount, setToolCount] = useState(0)
 
 	//TODO: to delete - not relevant
@@ -15,8 +15,15 @@ export default function StepFourTool({ formikProps }) {
 	// }
 
 	const {
+		data,
+		setData,
 		moveNextStep
-	} = useContext(SiteContext)
+	} = useContext(SiteContext);
+
+	const handleSubmit = (values, actions) => {
+		setData(prev => ({ ...prev, values }))
+		moveNextStep();
+	}
 
 	// TODO: State mutation
 	const onRemoveTool = (i) => {
@@ -145,29 +152,40 @@ export default function StepFourTool({ formikProps }) {
 
 	return (
 		<>
-			<div role="group">
-				<h4 style={{ margin: '10px 0' }}>האם קיימים ברשותך כלים שהיית רוצה להוריש?</h4>
-				<div className="status-group flex space-between input-btn">
-					<label className={`${formikProps.values.tool === "לא" ? 'active' : ''}`}>
-						<Field type="radio" name="tool" value="לא" />
-						לא
-					</label>
-					<label className={`${formikProps.values.tool === "כן" ? 'active' : ''}`}>
-						<Field type="radio" name="tool" value="כן" />
-						כן
-					</label>
-				</div>
-			</div>
-			{formikProps.values.tool === 'כן' &&
-				<div className="input-container-formik real-estate-container direction-ltr" >
-					<h4 className='direction-rtl' style={{ margin: '10px 0' }}>סוג הכלי</h4>
-					{renderToolsForm(formikProps)}
-				</div>
-			}
+			<Formik
+				//validationSchema={validationSchema}
+				initialValues={data}
+				onSubmit={handleSubmit}
+			>
+				{({ values }) => {
+					return (
+						<Form>
+							<div role="group">
+								<h4 style={{ margin: '10px 0' }}>האם קיימים ברשותך כלים שהיית רוצה להוריש?</h4>
+								<div className="status-group flex space-between input-btn">
+									<label className={`${values.tool === "לא" ? 'active' : ''}`}>
+										<Field type="radio" name="tool" value="לא" />
+										לא
+									</label>
+									<label className={`${values.tool === "כן" ? 'active' : ''}`}>
+										<Field type="radio" name="tool" value="כן" />
+										כן
+									</label>
+								</div>
+							</div>
+							{values.tool === 'כן' &&
+								<div className="input-container-formik real-estate-container direction-ltr" >
+									<h4 className='direction-rtl' style={{ margin: '10px 0' }}>סוג הכלי</h4>
+									{renderToolsForm(values)}
+								</div>
+							}
+							<button type="submit">המשך</button>
+						</Form>
+					)
+				}
+				}
+			</Formik>
 
-
-
-			<button onClick={() => moveNextStep()}>המשך</button>
 		</>
 	)
 }
